Added initial Gantt chart
This chart is an initial illustration and does not represent the final product. It is to be used as a simple guide to give a rough idea to when each part of the project will be developed. It will be used for comparison and feedback at the end of the project.